backend/requirements.txt
# FastAPI and web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.0
websockets>=12.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Background tasks
celery>=5.3.0
redis>=5.0.0

# Image processing
opencv-python-headless>=4.8.0
numpy>=1.26.0
scipy>=1.11.0
pillow>=10.1.0

# Visualization
matplotlib>=3.8.0

# RF Propagation (PyLayers - install from GitHub if needed)
# pip install git+https://github.com/pylayers/pylayers.git
shapely>=2.0.0
networkx>=3.2.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0

# AI/LLM
anthropic>=0.39.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
